+++ src/itest/java/com/iexec/blockchain/IntegrationTests.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2021-2023 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.iexec.blockchain;
 
 import com.iexec.blockchain.api.BlockchainAdapterApiClient;
@@ -6,19 +22,19 @@ import com.iexec.blockchain.broker.BrokerService;
 import com.iexec.blockchain.signer.SignerService;
 import com.iexec.blockchain.tool.CredentialsService;
 import com.iexec.blockchain.tool.IexecHubService;
-import com.iexec.common.chain.*;
 import com.iexec.common.chain.adapter.args.TaskContributeArgs;
 import com.iexec.common.chain.adapter.args.TaskFinalizeArgs;
 import com.iexec.common.chain.adapter.args.TaskRevealArgs;
 import com.iexec.common.sdk.broker.BrokerOrder;
-import com.iexec.common.sdk.order.payload.AppOrder;
-import com.iexec.common.sdk.order.payload.DatasetOrder;
-import com.iexec.common.sdk.order.payload.RequestOrder;
-import com.iexec.common.sdk.order.payload.WorkerpoolOrder;
-import com.iexec.common.security.Signature;
-import com.iexec.common.tee.TeeUtils;
-import com.iexec.common.utils.BytesUtils;
-import com.iexec.common.utils.HashUtils;
+import com.iexec.commons.poco.chain.*;
+import com.iexec.commons.poco.order.AppOrder;
+import com.iexec.commons.poco.order.DatasetOrder;
+import com.iexec.commons.poco.order.RequestOrder;
+import com.iexec.commons.poco.order.WorkerpoolOrder;
+import com.iexec.commons.poco.security.Signature;
+import com.iexec.commons.poco.tee.TeeUtils;
+import com.iexec.commons.poco.utils.BytesUtils;
+import com.iexec.commons.poco.utils.HashUtils;
 import feign.Logger;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -26,27 +42,38 @@ import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.DockerComposeContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 import org.web3j.crypto.Hash;
 import org.web3j.crypto.Sign;
 
+import java.io.File;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import java.util.stream.IntStream;
 
-import static com.iexec.common.chain.ChainTaskStatus.ACTIVE;
-import static com.iexec.common.chain.ChainTaskStatus.UNSET;
+import static com.iexec.commons.poco.chain.ChainTaskStatus.ACTIVE;
+import static com.iexec.commons.poco.chain.ChainTaskStatus.UNSET;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Testcontainers
 @ActiveProfiles("itest")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class IntegrationTests {
 
     public static final String USER = "admin";
@@ -55,8 +82,20 @@ class IntegrationTests {
     public static final int MAX_BLOCK_TO_WAIT = 3;
     public static final int POLLING_PER_BLOCK = 2;
     public static final int POLLING_INTERVAL_MS = BLOCK_TIME_MS / POLLING_PER_BLOCK;
-    public static final int MAX_POLLING_ATTEMPTS = MAX_BLOCK_TO_WAIT
-            * POLLING_PER_BLOCK;
+    public static final int MAX_POLLING_ATTEMPTS = MAX_BLOCK_TO_WAIT * POLLING_PER_BLOCK;
+
+    @Container
+    static DockerComposeContainer<?> environment = new DockerComposeContainer<>(new File("docker-compose.yml"))
+            .withExposedService("ibaa-chain", 8545, Wait.forListeningPort())
+            .withExposedService("ibaa-blockchain-adapter-mongo", 13012, Wait.forListeningPort());
+
+    @DynamicPropertySource
+    static void registerProperties(DynamicPropertyRegistry registry) {
+        registry.add("chain.id", () -> "65535");
+        registry.add("chain.hubAddress", () -> "0xC129e7917b7c7DeDfAa5Fff1FB18d5D7050fE8ca");
+        registry.add("chain.nodeAddress", () -> getServiceUrl(environment.getServicePort("ibaa-chain", 8545)));
+        registry.add("spring.data.mongodb.port", () -> environment.getServicePort("ibaa-blockchain-adapter-mongo", 13012));
+    }
 
     @LocalServerPort
     private int randomServerPort;
@@ -75,9 +114,15 @@ class IntegrationTests {
     private BlockchainAdapterApiClient appClient;
 
     @BeforeEach
-    void setUp() {
+    void setUp(TestInfo testInfo) {
+        log.info(">>> {}", testInfo.getDisplayName());
         appClient = BlockchainAdapterApiClientBuilder
-                .getInstanceWithBasicAuth(Logger.Level.FULL, "http://localhost:" + randomServerPort, USER, PASSWORD);
+                .getInstanceWithBasicAuth(Logger.Level.FULL, getServiceUrl(randomServerPort), USER, PASSWORD);
+    }
+
+    private static String getServiceUrl(int servicePort) {
+        log.info("service url http://localhost:{}", servicePort);
+        return "http://localhost:" + servicePort;
     }
 
     @Test
@@ -121,7 +166,7 @@ class IntegrationTests {
     }
 
     @Test
-    public void shouldBurstTransactionsWithAverageOfOneTxPerBlock(){
+    public void shouldBurstTransactionsWithAverageOfOneTxPerBlock() throws Exception {
         int taskVolume = 10;//small volume ensures reasonable execution time on CI/CD
         String dealId = triggerDeal(taskVolume);
         List<CompletableFuture<Void>> txCompletionWatchers = new ArrayList<>();
@@ -143,7 +188,7 @@ class IntegrationTests {
                             //no need to wait for propagation update in db
                             Assertions.assertTrue(true);
                         } catch (Exception e) {
-                            e.printStackTrace();
+                            log.error("Watcher failed with an exception", e);
                             Assertions.fail();
                         }
                     }));
@@ -151,7 +196,7 @@ class IntegrationTests {
         txCompletionWatchers.forEach(CompletableFuture::join);
     }
 
-    private String triggerDeal(int taskVolume) {
+    private String triggerDeal(int taskVolume) throws Exception {
         int secondsPollingInterval = POLLING_INTERVAL_MS / 1000;
         int secondsTimeout = secondsPollingInterval * MAX_POLLING_ATTEMPTS;
         String appAddress = iexecHubService.createApp(buildRandomName("app"),
@@ -192,10 +237,9 @@ class IntegrationTests {
         String dealId = brokerService.matchOrders(brokerOrder);
         Assertions.assertTrue(StringUtils.isNotEmpty(dealId));
         log.info("Created deal: {}", dealId);
-        //no need to wait since broker is synchronous, just checking deal
-        //existence for double checking
+        // no need to wait since broker is synchronous, just checking deal existence
         Optional<ChainDeal> chainDeal = iexecHubService.getChainDeal(dealId);
-        Assertions.assertTrue(chainDeal.isPresent());
+        assertThat(chainDeal).isPresent();
         return dealId;
     }
 
@@ -207,7 +251,7 @@ class IntegrationTests {
     private AppOrder buildAppOrder(String appAddress, int volume) {
         return AppOrder.builder()
                 .app(appAddress)
-                .price(BigInteger.ZERO)
+                .appprice(BigInteger.ZERO)
                 .volume(BigInteger.valueOf(volume))
                 .tag(BytesUtils.EMPTY_HEX_STRING_32)
                 .datasetrestrict(BytesUtils.EMPTY_ADDRESS)
@@ -220,7 +264,7 @@ class IntegrationTests {
     private WorkerpoolOrder buildWorkerpoolOrder(String workerpoolAddress, int volume) {
         return WorkerpoolOrder.builder()
                 .workerpool(workerpoolAddress)
-                .price(BigInteger.ZERO)
+                .workerpoolprice(BigInteger.ZERO)
                 .volume(BigInteger.valueOf(volume))
                 .tag(BytesUtils.EMPTY_HEX_STRING_32)
                 .trust(BigInteger.ZERO)
@@ -235,7 +279,7 @@ class IntegrationTests {
     private DatasetOrder buildDatasetOrder(String datasetAddress, int volume) {
         return DatasetOrder.builder()
                 .dataset(datasetAddress)
-                .price(BigInteger.ZERO)
+                .datasetprice(BigInteger.ZERO)
                 .volume(BigInteger.valueOf(volume))
                 .tag(BytesUtils.EMPTY_HEX_STRING_32)
                 .apprestrict(BytesUtils.EMPTY_ADDRESS)
@@ -287,18 +331,18 @@ class IntegrationTests {
         ChainTaskStatus status = null;
         int attempts = 0;
         while(true) {
-            log.info("Status [status:{}, chainTaskId:{}]", status, chainTaskId);
+            attempts++;
+            log.info("Status [status:{}, chainTaskId:{}, attempt:{}]", status, chainTaskId, attempts);
             status = iexecHubService.getChainTask(chainTaskId)
                     .map(ChainTask::getStatus)
                     .orElse(UNSET);
-            attempts++;
             if (status.equals(statusToWait) || attempts > maxAttempts) {
                 break;
             }
             TimeUnit.MILLISECONDS.sleep(pollingTimeMs);
         }
         if (!status.equals(statusToWait)) {
-            throw new Exception("Too long to wait for task: " + chainTaskId);
+            throw new TimeoutException("Too long to wait for task: " + chainTaskId);
         }
         log.info("Status reached [status:{}, chainTaskId:{}]", status, chainTaskId);
     }
@@ -312,31 +356,31 @@ class IntegrationTests {
         int winnerCounter = chainTask.getWinnerCounter();
         int revealCounter = chainTask.getRevealCounter();
         int attempts = 0;
+        log.info("{} {}", POLLING_INTERVAL_MS, MAX_POLLING_ATTEMPTS);
         while (revealCounter != winnerCounter) {
-            log.info("Waiting for reveals ({}/{})", revealCounter, winnerCounter);
-            Thread.sleep(POLLING_INTERVAL_MS);
+            attempts++;
+            log.info("Waiting for reveals ({}/{}), attempt {}", revealCounter, winnerCounter, attempts);
+            Thread.sleep(BLOCK_TIME_MS);
             revealCounter = iexecHubService.getChainTask(chainTaskId)
                     .map(ChainTask::getRevealCounter)
                     .orElse(0);
-            attempts++;
-            if (attempts == MAX_POLLING_ATTEMPTS) {
-                throw new Exception("Too long to wait for reveal: " + chainTaskId);
+            if (attempts == MAX_BLOCK_TO_WAIT) {
+                throw new TimeoutException("Too long to wait for reveal: " + chainTaskId);
             }
         }
-        System.out.println("All revealed (" + revealCounter + "/" + winnerCounter + ")");
+        log.info("All revealed ({}/{})", revealCounter, winnerCounter);
     }
 
     public WorkerpoolAuthorization mockAuthorization(String chainTaskId,
                                                      String enclaveChallenge) {
         String workerWallet = credentialsService.getCredentials().getAddress();
-        String hash =
-                HashUtils.concatenateAndHash(workerWallet,
-                        chainTaskId,
-                        enclaveChallenge);
+        String hash = HashUtils.concatenateAndHash(
+                workerWallet,
+                chainTaskId,
+                enclaveChallenge);
 
-        Sign.SignatureData sign =
-                Sign.signPrefixedMessage(BytesUtils.stringToBytes(hash),
-                        credentialsService.getCredentials().getEcKeyPair());
+        Sign.SignatureData sign = Sign.signPrefixedMessage(BytesUtils.stringToBytes(hash),
+                credentialsService.getCredentials().getEcKeyPair());
 
         return WorkerpoolAuthorization.builder()
                 .workerWallet(workerWallet)