+++ src/main/java/com/iexec/blockchain/command/task/initialize/TaskInitializeService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ package com.iexec.blockchain.command.task.initialize;
 
 import com.iexec.blockchain.command.generic.CommandEngine;
 import com.iexec.blockchain.tool.QueueService;
-import com.iexec.common.chain.ChainUtils;
+import com.iexec.commons.poco.chain.ChainUtils;
 import org.springframework.stereotype.Service;
 
 import static com.iexec.blockchain.tool.IexecHubService.isByte32;